работают сохранение и загрузка цвета для json

+++ GraphicEditor/Circle.cs
@@ -46,10 +46,10 @@ namespace GraphicEditor
             StrokeThickness = strokeThickness;
         }
         public bool IsSelected { get; set; }
-        public int Color { get; set; } = unchecked((int)0xFF000000);
+        public uint Color { get; set; } = unchecked((uint)0xFF000000);
         public void SetColor(byte a, byte r, byte g, byte b)
         {
-            Color = (a << 24) | (r << 16) | (g << 8) | b;
+            Color = (uint)((a << 24) | (r << 16) | (g << 8) | b);
         }
 
         public void Move(Point vector)
