+++ GraphicEditor/Triangle.cs
@@ -59,10 +59,10 @@ namespace GraphicEditor
         }
         public bool IsSelected { get; set; }
         public double StrokeThickness { get; set; } = 2;
-        public int Color { get; set; } = unchecked((int)0xFF000000);
+        public uint Color { get; set; } = unchecked((uint)0xFF000000);
         public void SetColor(byte a, byte r, byte g, byte b)
         {
-            Color = (a << 24) | (r << 16) | (g << 8) | b;
+            Color = (uint)((a << 24) | (r << 16) | (g << 8) | b);
         }
         public Point Center => new Point((P1.X + P2.X + P3.X) / 3, (P1.Y + P2.Y + P3.Y) / 3 );
